corrige formulário e cria callbacks necessários - criação de locação

+++ app/models/price.rb
@@ -2,5 +2,5 @@ class Price
   include Mongoid::Document
 
   field :per_km, type: Float
-  field :percentage, type: Float
+  field :percentage, type: Integer
 end
